Reproduced error for bankai issue https://github.com/choojs/bankai/issues/495
