+++ index.dump.js
@@ -1,0 +1,20 @@
+var css = require('sheetify')
+var choo = require('choo')
+
+css('tachyons')
+css('./main.css')
+css('./main.extend.css')
+
+var app = choo()
+if (process.env.NODE_ENV !== 'production') {
+  app.use(require('choo-devtools')())
+} else {
+  app.use(require('choo-service-worker')())
+}
+
+app.use(require('./stores/clicks'))
+
+app.route('/', require('./views/main'))
+app.route('/*', require('./views/404'))
+
+module.exports = app.mount('body')
