+++ index.js
@@ -1,19 +1,18 @@
-var css = require('sheetify')
 var choo = require('choo')
+var css = require('sheetify')
+
+var MainView = require('./views/main')
 
 css('tachyons')
-css('./main.css')
+css('./tachyons.extend.css')
+css('./fonts.css')
 
-var app = choo()
+const app = choo()
 if (process.env.NODE_ENV !== 'production') {
   app.use(require('choo-devtools')())
 } else {
+  app.use(require('choo-service-worker/clear')())
   app.use(require('choo-service-worker')())
 }
-
-app.use(require('./stores/clicks'))
-
-app.route('/', require('./views/main'))
-app.route('/*', require('./views/404'))
-
-module.exports = app.mount('body')
+app.route('/', MainView)
+app.mount('body')
