--- main.css
@@ -1,3 +1,0 @@
-h2{
-  color: green;
-}
